+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * add more Documentation
 * add ToUnixTimeStamp at DateTime
 * Reactive Store emits Subscription only when the value was changed
+* add Pull Function to Array
 
 ### 3.0.0